Fix title booting
We achieve this by not reloading IOS.

If reloaded prior to ES_Identify, ES returned a title ID of 0. This breaks the NAND boot program, as it expects to find its title ID and load the actual binary.

While one could most likely reload IOS and selectively apply patches, it seems like a logistical nightmare with this current setup. I am not touching it again.

Cheap hack? Probably. Usable for our needs? We'll see. (This method seems to work so far, at least.)