Insert stub after NAND Boot Program
We load the NAND Boot Program normally.

The NAND Boot Program appears to handle tasks such as setting the proper video environment, inserting exceptions and clearing state. It's unknown as to why it's needed - channels do this anyway. Oh well, it's no issue, and allows us to use a LZ11
compressed DOL should we desire. (We do. It saves nearly 4 mb.)

Once loaded, we patch its operation after the end of OSInit to instead jump to our stub, which as of this commit simply does the same as OSInit would normally. We will later modify it to properly patch.